Make changes in the codebase to crap code for elk configuration and sync logstash process 	modified:   browserium/driver.py 	deleted:    browserium/elkStackConfigurer/darwin/__init__.py 	deleted:    browserium/elkStackConfigurer/darwin/kibana.yml 	deleted:    browserium/elkStackConfigurer/darwin/mac_installation.sh 	deleted:    browserium/elkStackConfigurer/debian/__init__.py 	deleted:    browserium/elkStackConfigurer/debian/debian_installation.sh 	deleted:    browserium/elkStackConfigurer/debian/kibana.yml 	deleted:    browserium/elkStackConfigurer/elk_configure.py 	deleted:    browserium/elkStackConfigurer/rpm/__init__.py 	deleted:    browserium/elkStackConfigurer/rpm/elasticsearch.repo 	deleted:    browserium/elkStackConfigurer/rpm/kibana.repo 	deleted:    browserium/elkStackConfigurer/rpm/kibana.yml 	deleted:    browserium/elkStackConfigurer/rpm/logstash.repo 	deleted:    browserium/elkStackConfigurer/rpm/rpm_installation.sh 	renamed:    browserium/elkStackConfigurer/__init__.py -> browserium/logstashRunner/__init__.py 	renamed:    browserium/elkStackConfigurer/logstash.conf -> browserium/logstashRunner/logstash.conf 	new file:   browserium/logstashRunner/logstash_runner.py 	deleted:    browserium/utility/distType.sh 	modified:   browserium/utility/main.py 	modified:   browserium/utility/os_type.py 	modified:   setup.py
