--- browserium/utility/distType.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-cat /etc/*release | grep 'ID_LIKE'
