+++ browserium/utility/main.py
@@ -8,12 +8,12 @@ subparser = parser.add_subparsers(dest='command',help="download")
 download = subparser.add_parser("download", help="download driver binary")
 update = subparser.add_parser("update", help="update driver binary")
 delete = subparser.add_parser("delete", help="delete specific driver binary")
-elk = subparser.add_parser("elk", help="install elk stack")
+logstash = subparser.add_parser("logstash", help="run logstash in the background")
 
 download.add_argument("--driver", help="download the respective driver")
 update.add_argument("--driver", help="update the respective driver")
 delete.add_argument("--driver", help="delete all drivers")
-elk.add_argument("--driver", help="install elk stack")
+logstash.add_argument("--driver", help="run logstash in the background")
 
 def main():
     args = parser.parse_args()
@@ -24,7 +24,7 @@ def main():
         driver_object.update_driver()
     elif args.command == "delete":
         driver_object.delete_driver_history()
-    elif args.command == "elk":
+    elif args.command == "logstash":
         driver_object.Elkmain()
 
 if __name__ == '__main__':
